* Separated test scenes into different files * The window now remembers the last rendered image and will display it even after minimizing * Added max distance parameter to Bvh intersection test * Attenuation now returns floats (but is currently not really used anyways) * Added useful color functions like clamp and mix * Color-scalar operations don't use the quadrat of the scalar anymore * Fixed error in randomHemisphere method on Vec3d * Added weighted hemisphere method on Vec3d for importance sampling * Added random normal function to Vec3d for fog scattering * Changed the material trait to support sub surface scattering and light absorbtion * Added sky and air material to the scene * Added inside out intersection to AABBs * Removed direct lighting * Added volumetric scattering and distance based absorbtion to the tracing worker

+++ Code/SPath/src/kaloffl/spath/math/Attenuation.scala
@@ -12,7 +12,7 @@ class Attenuation(
     val linear: Double,
     val quadratic: Double) {
 
-  def apply(dist: Double): Double = {
-    return 1.0 / (constant + linear * dist + quadratic * dist * dist)
+  def apply(dist: Double): Float = {
+    return 1.0f / (constant + linear * dist + quadratic * dist * dist).toFloat
   }
 }
